requirements.txt
playwright==1.42.0
requests==2.31.0
python-dotenv==1.0.0
anthropic==0.25.0
